backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.8
httpx==0.25.2
psutil==5.9.6
PySocks==1.7.1